% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycode.R
\name{powerproduction}
\alias{powerproduction}
\title{powerproduction}
\usage{
powerproduction(.data, name_power_station)
}
\value{
Returning yearly power production per power station
}
\description{
powerproduction is a function that uses data from the use_api() function to load data from about Norwegian hydro power from Norges vassdrags- og energidirektorat (NVE), and then returns a data frame showing power stations and yearly power production for that power station.
}
\examples{
powerproduction
data <- powerproduction() returns a data frame with 1672 oberservation of two variables, MidProd_81_10 and Navn

}
