% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycode.R
\name{powerproduction_for_one}
\alias{powerproduction_for_one}
\title{power_production_for_one}
\usage{
powerproduction_for_one(.data, name_power_station)
}
\value{
Returning yearly power production per power station
}
\description{
powerproduction_for_one is a function that uses data from the use_api() function to load data from about Norwegian hydro power from Norges vassdrags- og energidirektorat (NVE), and then returns a data frame showing a selected power station and its yearly power production.
}
\details{
powerproduction_for_one
}
\examples{
grondal <- powerproduction_for_one(name_power_station = "Grøndal") returns a data frame with 1 oberservation of two variables, MidProd_81_10 and Navn. This is the yearly power production for the power station, Grøndal.

}
