% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycode.R
\name{use_api}
\alias{use_api}
\title{use_api}
\usage{
use_api()
}
\value{
Loading data from api
}
\description{
use_api loads data from about Norwegian hydro power from Norges vassdrags- og energidirektorat (NVE) and saves it to a data frame.
}
